Refs #24793 - split service classes into separate files
