+++ lib/foreman_maintain/utils.rb
@@ -4,6 +4,6 @@ require 'foreman_maintain/utils/bash'
 require 'foreman_maintain/utils/hash_tools'
 require 'foreman_maintain/utils/curl_response'
 require 'foreman_maintain/utils/mongo_core'
-require 'foreman_maintain/utils/system_service'
+require 'foreman_maintain/utils/service'
 require 'foreman_maintain/utils/system_helpers'
 require 'foreman_maintain/utils/facter'
