+++ lib/foreman_maintain/utils/service.rb
@@ -1,0 +1,14 @@
+require 'foreman_maintain/utils/service/abstract'
+require 'foreman_maintain/utils/service/systemd'
+require 'foreman_maintain/utils/service/remote_db'
+
+module ForemanMaintain::Utils
+  def self.system_service(name, priority, options = {})
+    db_feature = options.fetch(:db_feature, nil)
+    if name =~ /^(postgresql|.*mongod)$/ && db_feature && db_feature.local? == false
+      Service::RemoteDB.new(name, priority, options)
+    else
+      Service::Systemd.new(name, priority, options)
+    end
+  end
+end
